miku_assistant/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
livekit
livekit-api
openai
elevenlabs
sqlalchemy
requests
python-multipart
websockets
asyncio
